pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "alertmcp"
version = "0.1.0"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
description = "デスクトップ通知機能を提供するMCPツール"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastmcp>=2.2.0",
    "plyer>=2.1.0",
]

[project.optional-dependencies]
macos = ["pyobjus>=1.3.0"]

[project.urls]
"Homepage" = "https://github.com/kousunh/alertmcp"
"Bug Tracker" = "https://github.com/kousunh/alertmcp/issues"

[tool.setuptools]
packages = ["alertmcp"] 